requirements.txt
altgraph==0.17.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
future==0.18.2
idna==3.3
pefile==2021.9.3
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
pyserial==3.5
python-dotenv==0.19.2
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
urllib3==1.26.9
